Remove support for Ionic 3
