+++ generators/entity-client/index.js
@@ -18,6 +18,7 @@
  */
 /* eslint-disable consistent-return */
 const chalk = require('chalk');
+const _ = require('lodash');
 const utils = require('generator-jhipster/generators/utils');
 const BaseGenerator = require('generator-jhipster/generators/generator-base');
 const writeFiles = require('./files').writeFiles;
@@ -39,7 +40,7 @@ module.exports = class extends BaseGenerator {
 
   end() {
     if (useBlueprint) return;
-    this.log(chalk.bold.green('Entity generation completed. You may have to stop/start "ionic serve" to see your new entity.'));
+    this.log(chalk.bold.green('\nEntity generation complete!'));
   }
 
   /**
@@ -57,17 +58,70 @@ module.exports = class extends BaseGenerator {
     if (!entityAngularName) {
       return;
     }
-    const entityPagePath = 'src/pages/entities/entity.ts';
+    const entityPagePath = 'src/app/pages/entities/entities.page.ts';
     try {
-      const page = `{name: '${entityAngularName}', component: '${entityAngularName}Page'},`;
-      utils.rewriteFile(
-        {
-          file: entityPagePath,
-          needle: 'jhipster-needle-add-entity-page',
-          splicable: [this.stripMargin(page)]
-        },
-        this
+      const isSpecificEntityAlreadyGenerated = utils.checkStringInFile(entityPagePath, `route: '${entityFileName}'`, this);
+
+      if (!isSpecificEntityAlreadyGenerated) {
+        const isAnyEntityAlreadyGenerated = utils.checkStringInFile(entityPagePath, 'route:', this);
+        const prefix = isAnyEntityAlreadyGenerated ? ',' : '';
+        const pageEntry = `${prefix}{ name: '${entityAngularName}', component: '${entityAngularName}Page', route: '${entityFileName}' }`;
+        utils.rewriteFile(
+          {
+            file: entityPagePath,
+            needle: 'jhipster-needle-add-entity-page',
+            splicable: [this.stripMargin(pageEntry)]
+          },
+          this
+        );
+      }
+    } catch (e) {
+      this.log(
+        `${
+          chalk.yellow('\nUnable to find ') +
+          entityPagePath +
+          chalk.yellow(' or missing required jhipster-needle. Reference to ') +
+          entityAngularName
+        } ${chalk.yellow(`not added to ${entityPagePath}.\n`)}`
       );
+      this.debug('Error:', e);
+    }
+  }
+
+  /**
+   * Add a new route in the TS modules file.
+   *
+   * @param {string} entityInstance - Entity Instance
+   * @param {string} entityClass - Entity Class
+   * @param {string} entityAngularName - Entity Angular Name
+   * @param {string} entityFolderName - Entity Folder Name
+   * @param {string} entityFileName - Entity File Name
+   * @param {boolean} enableTranslation - If translations are enabled or not
+   */
+  addEntityRouteToModule(entityInstance, entityClass, entityAngularName, entityFolderName, entityFileName, enableTranslation) {
+    // workaround method being called on initialization
+    if (!entityAngularName) {
+      return;
+    }
+    const entityPagePath = 'src/app/pages/entities/entities.module.ts';
+    try {
+      const isSpecificEntityAlreadyGenerated = utils.checkStringInFile(entityPagePath, `path: '${entityFileName}'`, this);
+      if (!isSpecificEntityAlreadyGenerated) {
+        const isAnyEntityAlreadyGenerated = utils.checkStringInFile(entityPagePath, 'loadChildren', this);
+        const prefix = isAnyEntityAlreadyGenerated ? ',' : '';
+        const route = `|${prefix} {
+                    |    path: '${entityFileName}',
+                    |    loadChildren: './${entityFolderName}/${entityFileName}.module#${entityAngularName}PageModule'
+                    |  }`;
+        utils.rewriteFile(
+          {
+            file: entityPagePath,
+            needle: 'jhipster-needle-add-entity-route',
+            splicable: [this.stripMargin(route)]
+          },
+          this
+        );
+      }
     } catch (e) {
       this.log(
         `${
@@ -102,7 +156,7 @@ module.exports = class extends BaseGenerator {
       let variableName;
       hasManyToMany = hasManyToMany || relationship.relationshipType === 'many-to-many';
       if (relationship.relationshipType === 'one-to-one' && relationship.ownerSide === true && relationship.otherEntityName !== 'user') {
-        variableName = relationship.relationshipFieldNamePlural.toLowerCase();
+        variableName = _.camelCase(relationship.otherEntityNameCapitalizedPlural);
         if (variableName === entityInstance) {
           variableName += 'Collection';
         }
@@ -114,22 +168,22 @@ module.exports = class extends BaseGenerator {
             .query({filter: '${relationship.otherEntityRelationshipName.toLowerCase()}-is-null'})
             .subscribe(data => {
                 if (${relationshipFieldNameIdCheck}) {
-                    this.${variableName} = data;
+                    this.${variableName} = data.body;
                 } else {
                     this.${relationship.otherEntityName}Service
                         .find(${relationshipFieldName}${dto === 'no' ? '.id' : 'Id'})
-                        .subscribe((subData: ${relationship.otherEntityAngularName}) => {
-                            this.${variableName} = [subData].concat(subData);
+                        .subscribe((subData: HttpResponse<${relationship.otherEntityAngularName}>) => {
+                            this.${variableName} = [subData.body].concat(subData.body);
                         }, (error) => this.onError(error));
                 }
             }, (error) => this.onError(error));`;
       } else if (relationship.relationshipType !== 'one-to-many') {
-        variableName = relationship.otherEntityNameCapitalizedPlural.toLowerCase();
+        variableName = _.camelCase(relationship.otherEntityNameCapitalizedPlural);
         if (variableName === entityInstance) {
           variableName += 'Collection';
         }
         query = `this.${relationship.otherEntityName}Service.query()
-            .subscribe(data => { this.${variableName} = data; }, (error) => this.onError(error));`;
+            .subscribe(data => { this.${variableName} = data.body; }, (error) => this.onError(error));`;
       }
       if (variableName && !this.contains(queries, query)) {
         queries.push(query);
