+++ generators/entity-client/templates/ionic/src/app/pages/entities/_entity-update.spec.ts.ejs
