+++ generators/entity-client/templates/ionic/src/app/pages/entities/_entity.service.ts.ejs
