+++ generators/entity-client/templates/ionic/src/app/pages/entities/_entity.spec.ts.ejs
