+++ generators/entity-client/templates/ionic/src/app/pages/entities/_index.ts.ejs
