+++ generators/entity-client/templates/ionic/src/app/pages/entities/infinite-scroll-template.ejs
