+++ generators/entity-client/templates/ionic/src/app/pages/entities/ng_validators.ejs
