+++ generators/entity-client/templates/ionic/src/app/pages/entities/no-pagination-template.ejs
