+++ generators/entity-client/templates/ionic/src/app/pages/entities/pagination-template.ejs
