--- generators/entity-client/templates/ionic/src/pages/entities/_entity-detail.html.ejs
@@ -1,113 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-<%_ const keyPrefix = entityTranslationKey; _%>
-<ion-header>
-    <ion-navbar>
-        <ion-title><%= entityClassHumanized %></ion-title>
-    </ion-navbar>
-</ion-header>
-<ion-content class="ion-padding">
-    <ion-list>
-        <%_ for (idx in fields) {
-        const fieldName = fields[idx].fieldName;
-        const fieldType = fields[idx].fieldType;
-        const fieldTypeBlobContent = fields[idx].fieldTypeBlobContent; _%>
-        <ion-item>
-            <ion-label fixed><%= fields[idx].fieldNameHumanized %></ion-label>
-            <div item-content>
-                <%_ if (fields[idx].fieldIsEnum) { _%>
-                <span>{{<%= entityInstance %>.<%= fieldName %>}}</span>
-                <%_ } else if (['byte[]', 'ByteBuffer'].includes(fieldType) && fieldTypeBlobContent === 'image') { _%>
-                <div *ngIf="<%= entityInstance %>.<%= fieldName %>">
-                    <a (click)="openFile(<%= entityInstance %>.<%= fieldName %>ContentType, <%= entityInstance %>.<%= fieldName %>)">
-                        <img [src]="'data:' + <%=entityInstance %>.<%= fieldName %>ContentType + ';base64,' + <%= entityInstance %>.<%= fieldName %>" style="max-width: 100%;" alt="<%=entityInstance %> image"/>
-                    </a>
-                    {{<%= entityInstance %>.<%= fieldName %>ContentType}}, {{byteSize(<%= entityInstance %>.<%= fieldName %>)}}
-                </div>
-                <%_ } else if (['Instant', 'ZonedDateTime'].includes(fieldType)) { _%>
-                <span>{{<%=entityInstance %>.<%= fieldName %> | date:'medium'}}</span>
-                <%_ } else if (fieldType === 'LocalDate') { _%>
-                <span>{{<%=entityInstance %>.<%= fieldName %> | date:'mediumDate'}}</span>
-                <%_ } else if (['byte[]', 'ByteBuffer'].includes(fieldType) && fieldTypeBlobContent === 'any') { _%>
-                <div *ngIf="<%= entityInstance %>.<%= fieldName %>">
-                    <a (click)="openFile(<%= entityInstance %>.<%= fieldName %>ContentType, <%= entityInstance %>.<%= fieldName %>)" jhiTranslate="entity.action.open">open</a>
-                    {{<%= entityInstance %>.<%= fieldName %>ContentType}}, {{byteSize(<%= entityInstance %>.<%= fieldName %>)}}
-                </div>
-                <%_ } else { _%>
-                <span>{{<%= entityInstance %>.<%= fieldName %>}}</span>
-                <%_ } _%>
-            </div>
-        </ion-item>
-        <%_ } _%>
-        <%_ for (idx in relationships) {
-        const relationshipType = relationships[idx].relationshipType;
-        const ownerSide = relationships[idx].ownerSide;
-        const relationshipName = relationships[idx].relationshipName;
-        const relationshipFieldName = relationships[idx].relationshipFieldName;
-        const relationshipFieldNamePlural = relationships[idx].relationshipFieldNamePlural;
-        const relationshipNameHumanized = relationships[idx].relationshipNameHumanized;
-        const otherEntityName = relationships[idx].otherEntityName;
-        const otherEntityStateName = relationships[idx].otherEntityStateName;
-        const otherEntityField = relationships[idx].otherEntityField;
-        const otherEntityFieldCapitalized = relationships[idx].otherEntityFieldCapitalized;
-        if (relationshipType === 'many-to-one'
-        || (relationshipType === 'one-to-one' && ownerSide === true)
-        || (relationshipType === 'many-to-many' && ownerSide === true)) { _%>
-        <ion-item>
-            <ion-label><%= relationshipNameHumanized %></ion-label>
-            <%_ if (otherEntityName === 'user') { _%>
-                <%_ if (relationshipType === 'many-to-many') { _%>
-            <div item-content>
-                <span *ngFor="let <%= relationshipFieldName %> of <%= entityInstance %>.<%= relationshipFieldNamePlural %>; let last = last">
-                    {{<%= relationshipFieldName %>.<%= otherEntityField %>}}{{last ? '' : ', '}}
-                </span>
-            </div>
-                <%_ } else { _%>
-                    <%_ if (dto === 'no') { _%>
-            <div item-content>{{<%= entityInstance + "." + relationshipFieldName + "?." + otherEntityField %>}}</div>
-                    <%_ } else { _%>
-            <div item-content>{{<%= entityInstance + "." + relationshipFieldName + otherEntityFieldCapitalized %>}}</div>
-                    <%_ } _%>
-                <%_ } _%>
-            <%_ } else { _%>
-                <%_ if (relationshipType === 'many-to-many') { _%>
-            <div item-content>
-                <span *ngFor="let <%= relationshipFieldName %> of <%= entityInstance %>.<%= relationshipFieldNamePlural %>; let last = last">
-                    <a>{{<%= relationshipFieldName %>.<%= otherEntityField %>}}</a>{{last ? '' : ', '}}
-                </span>
-            </div>
-                <%_ } else { _%>
-                    <%_ if (dto === 'no') { _%>
-            <div item-content *ngIf="<%= entityInstance + '.' + relationshipFieldName %>">
-                <a>{{<%= entityInstance + "." + relationshipFieldName + "?." + otherEntityField %>}}</a>
-            </div>
-                    <%_ } else { _%>
-            <div item-content *ngIf="<%= entityInstance + '.' + relationshipFieldName + "Id" %>">
-                <a>{{<%= entityInstance + "." + relationshipFieldName + otherEntityFieldCapitalized %>}}</a>
-            </div>
-                    <%_ } _%>
-                <%_ } _%>
-            <%_ } _%>
-        </ion-item>
-        <%_ } _%>
-        <%_ } _%>
-    </ion-list>
-
-    <button ion-button block (click)="open(<%=entityInstance %>)">{{ 'EDIT_BUTTON' | translate }}</button>
-</ion-content>
