--- generators/entity-client/templates/ionic/src/pages/entities/_entity-dialog.module.ts.ejs
@@ -1,69 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
-    This file is part of the JHipster project, see http://www.jhipster.tech/
-    for more information.
-
-        Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-limitations under the License.
--%>
-<%_
-let hasRelationshipQuery = false;
-Object.keys(differentRelationships).forEach(key => {
-    const hasAnyRelationshipQuery = differentRelationships[key].some(rel =>
-        (rel.relationshipType === 'one-to-one' && rel.ownerSide === true && rel.otherEntityName !== 'user')
-        || rel.relationshipType !== 'one-to-many'
-    );
-    if (hasAnyRelationshipQuery) {
-        hasRelationshipQuery = true;
-    }
-    if (differentRelationships[key].some(rel => rel.relationshipType !== 'one-to-many')) {
-        const uniqueRel = differentRelationships[key][0];
-        if (uniqueRel.otherEntityAngularName === 'User') { _%>
-import { User as UserService } from '../../../providers/user/user';
-            <%_ } else if (uniqueRel.otherEntityAngularName !== entityAngularName) { _%>
-import { <%= uniqueRel.otherEntityAngularName%>Service } from '../<%= uniqueRel.otherEntityModulePath %>';
-            <%_}
-    }
-}); _%>
-import { NgModule } from '@angular/core';
-import { TranslateModule } from '@ngx-translate/core';
-import { IonicPageModule } from 'ionic-angular';
-import { <%= entityAngularName %>DialogPage } from './<%= entityFileName %>-dialog';
-import { <%= entityAngularName %>Service } from './<%= entityFileName %>.provider';
-
-@NgModule({
-    declarations: [
-        <%= entityAngularName %>DialogPage
-    ],
-    imports: [
-        IonicPageModule.forChild(<%= entityAngularName %>DialogPage),
-        TranslateModule.forChild()
-    ],
-    exports: [
-        <%= entityAngularName %>DialogPage
-    ],
-    providers: [
-        <%= entityAngularName %>Service<% if (hasRelationshipQuery) { %>,<% } %>
-        <%_ Object.keys(differentRelationships).forEach(key => {
-            if (differentRelationships[key].some(rel => rel.relationshipType !== 'one-to-many')) {
-                const uniqueRel = differentRelationships[key][0];
-                if (uniqueRel.otherEntityAngularName !== entityAngularName) { _%>
-        <%= uniqueRel.otherEntityAngularName %>Service,
-             <%_}
-            }
-        });
-_%>
-    ]
-})
-export class <%= entityAngularName %>DialogPageModule {
-}
