--- generators/entity-client/templates/ionic/src/pages/entities/_entity.html.ejs
@@ -1,77 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-<%_ const keyPrefix = entityTranslationKey; _%>
-<ion-header>
-    <ion-navbar>
-        <ion-title><%= entityClassPluralHumanized %></ion-title>
-        <ion-buttons end>
-            <button ion-button icon-only (click)="open()">
-                <ion-icon name="add"></ion-icon>
-            </button>
-        </ion-buttons>
-    </ion-navbar>
-</ion-header>
-<!-- todo: add elasticsearch support -->
-<ion-content class="ion-padding">
-    <ion-refresher (ionRefresh)="loadAll($event)">
-        <ion-refresher-content></ion-refresher-content>
-    </ion-refresher>
-
-    <ion-list>
-        <ion-item-sliding *ngFor="let <%=entityInstance %> of <%=entityInstancePlural %>; trackBy: trackId" #slidingItem>
-            <button ion-item (click)="detail(<%= entityInstance %>)">
-              <%_ for (idx in fields) {
-                const fieldName = fields[idx].fieldName;
-                const fieldNameCapitalized = fields[idx].fieldNameCapitalized;
-                const fieldType = fields[idx].fieldType;
-                const fieldTypeBlobContent = fields[idx].fieldTypeBlobContent; _%>
-                <%_ if (['byte[]', 'ByteBuffer'].includes(fieldType) && fieldTypeBlobContent === 'image') { _%>
-                <ion-avatar<% if (idx === 0) { %>item-start<% } %> *ngIf="<%= entityInstance %>.<%= fieldName %>">
-                    <img [src]="'data:' + <%=entityInstance %>.<%=fieldName%>ContentType + ';base64,' + <%=entityInstance %>.<%=fieldName%>" />
-                </ion-avatar>
-                <%_ } else if (fields[idx].fieldIsEnum) { _%>
-                <!-- todo: special handling for translating enum - {{'<%= fieldType %>.' + <%= entityInstance %>.<%= fieldName %>}}" -->
-                <p>{{<%=entityInstance %>.<%=fieldName%>}}</p>
-                <%_ } else if (['Instant', 'ZonedDateTime'].includes(fieldType)) { _%>
-                <p>{{<%=entityInstance %>.<%=fieldName%> | date:'medium'}}</p>
-                <%_ } else if (fieldType === 'LocalDate') { _%>
-                <p>{{<%=entityInstance %>.<%=fieldName%> | date:'mediumDate'}}</p>
-                <%_ } else { _%>
-                  <%_ if (fieldName === 'name' || fieldName === 'title') { _%>
-                <h2>{{<%=entityInstance %>.<%=fieldName%>}}</h2>
-                  <%_ } else { _%>
-                <p>{{<%=entityInstance %>.<%=fieldName%>}}</p>
-                  <%_ } _%>
-                <%_ } _%>
-              <%_ } _%>
-            </button>
-            <ion-item-options>
-                <button ion-button color="primary" (click)="open(slidingItem, <%=entityInstance %>)">
-                    {{ 'EDIT_BUTTON' | translate }}
-                </button>
-                <button ion-button color="danger" (click)="delete(<%=entityInstance %>)">
-                    {{ 'DELETE_BUTTON' | translate }}
-                </button>
-            </ion-item-options>
-        </ion-item-sliding>
-    </ion-list>
-    <ion-item *ngIf="!<%=entityInstancePlural %>?.length">
-        No <%= entityClassPluralHumanized %> found.
-    </ion-item>
-</ion-content>
