--- generators/entity-client/templates/ionic/src/pages/entities/_entity.module.ts.ejs
@@ -1,39 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-import { NgModule } from '@angular/core';
-import { TranslateModule } from '@ngx-translate/core';
-import { IonicPageModule } from 'ionic-angular';
-import { <%= entityAngularName %>Page } from './<%=entityFileName %>';
-import { <%= entityAngularName %>Service } from './<%= entityFileName %>.provider';
-
-@NgModule({
-    declarations: [
-        <%= entityAngularName %>Page
-    ],
-    imports: [
-        IonicPageModule.forChild(<%= entityAngularName %>Page),
-        TranslateModule.forChild()
-    ],
-    exports: [
-        <%= entityAngularName %>Page
-    ],
-    providers: [<%= entityAngularName %>Service]
-})
-export class <%= entityAngularName %>PageModule {
-}
