--- generators/entity-client/templates/ionic/src/pages/entities/_index.ts.ejs
@@ -1,23 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-export * from './<%= entityFileName %>.model';
-export * from './<%= entityFileName %>.provider';
-export * from './<%= entityFileName %>-dialog';
-export * from './<%= entityFileName %>-detail';
-export * from './<%= entityFileName %>';
