--- generators/entity-client4/templates/ionic/src/app/pages/entities/infinite-scroll-template.ejs
@@ -1,131 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-<%_
-const i18nToLoad = [entityInstance];
-for (const idx in fields) {
-    if (fields[idx].fieldIsEnum === true) {
-        i18nToLoad.push(fields[idx].enumInstance);
-    }
-}
-_%>
-
-    <%= entityInstancePlural %>: <%= entityAngularName %>[];
-    currentAccount: any;
-    eventSubscriber: Subscription;
-    itemsPerPage: number;
-    links: any;
-    page: any;
-    predicate: any;
-    queryCount: any;
-    reverse: any;
-    totalItems: number;
-    <%_ if (searchEngine === 'elasticsearch') { _%>
-    currentSearch: string;
-    <%_ } _%>
-
-    constructor(
-        private <%= entityInstance %>Service: <%= entityAngularName %>Service,
-        private jhiAlertService: JhiAlertService,
-        <%_ if (fieldsContainBlob) { _%>
-        private dataUtils: JhiDataUtils,
-        <%_ } _%>
-        private eventManager: JhiEventManager,
-        private parseLinks: JhiParseLinks,
-        <%_ if (searchEngine === 'elasticsearch') { _%>
-        private activatedRoute: ActivatedRoute,
-        <%_ } _%>
-        private principal: Principal
-    ) {
-        this.<%= entityInstancePlural %> = [];
-        this.itemsPerPage = ITEMS_PER_PAGE;
-        this.page = 0;
-        this.links = {
-            last: 0
-        };
-        this.predicate = 'id';
-        this.reverse = true;
-        <%_ if (searchEngine === 'elasticsearch') { _%>
-        this.currentSearch = activatedRoute.snapshot.params['search'] ? activatedRoute.snapshot.params['search'] : '';
-        <%_ } _%>
-    }
-
-    loadAll() {
-    <%_ if (searchEngine === 'elasticsearch') { _%>
-        if (this.currentSearch) {
-            this.<%= entityInstance %>Service.search({
-                query: this.currentSearch,
-                page: this.page,
-                size: this.itemsPerPage,
-                sort: this.sort()
-            }).subscribe(
-                (res: ResponseWrapper) => this.onSuccess(res.json, res.headers),
-                (res: ResponseWrapper) => this.onError(res.json)
-            );
-            return;
-        }
-    <%_ } _%>
-        this.<%= entityInstance %>Service.query({
-            page: this.page,
-            size: this.itemsPerPage,
-            sort: this.sort()
-        }).subscribe(
-            (res: ResponseWrapper) => this.onSuccess(res.json, res.headers),
-            (res: ResponseWrapper) => this.onError(res.json)
-        );
-    }
-
-    reset() {
-        this.page = 0;
-        this.<%= entityInstancePlural %> = [];
-        this.loadAll();
-    }
-
-    loadPage(page) {
-        this.page = page;
-        this.loadAll();
-    }
-    <%_ if (searchEngine === 'elasticsearch') { _%>
-
-    clear() {
-        this.<%= entityInstancePlural %> = [];
-        this.links = {
-            last: 0
-        };
-        this.page = 0;
-        this.predicate = 'id';
-        this.reverse = true;
-        this.currentSearch = '';
-        this.loadAll();
-    }
-
-    search(query) {
-        if (!query) {
-            return this.clear();
-        }
-        this.<%= entityInstancePlural %> = [];
-        this.links = {
-            last: 0
-        };
-        this.page = 0;
-        this.predicate = '_score';
-        this.reverse = false;
-        this.currentSearch = query;
-        this.loadAll();
-    }
-    <%_ } _%>
