--- generators/entity-client4/templates/ionic/src/app/pages/entities/ng_validators.ejs
@@ -1,56 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-<%#	Returns a string of all angular validators required for the input field. -%>
-<%
-const field = fields[idx];
-let result = '';
-
-if (field.fieldValidate === true) {
-	const rules = field.fieldValidateRules;
-	const validators = [];
-
-    if (rules.indexOf('required') !== -1 && field.fieldType !== 'Boolean') {
-        validators.push('required');
-    }
-    if (rules.indexOf('minlength') !== -1) {
-        validators.push('minlength="' + field.fieldValidateRulesMinlength + '"');
-    }
-    if (rules.indexOf('maxlength') !== -1) {
-        validators.push('maxlength="' + field.fieldValidateRulesMaxlength + '"');
-    }
-    if (rules.indexOf('min') !== -1) {
-        validators.push('min="' + field.fieldValidateRulesMin + '"');
-    }
-    if (rules.indexOf('max') !== -1) {
-        validators.push('max="' + field.fieldValidateRulesMax + '"');
-    }
-    if (rules.indexOf('minbytes') !== -1) {
-        validators.push('minbytes="' + field.fieldValidateRulesMinbytes +'"');
-    }
-    if (rules.indexOf('maxbytes') !== -1) {
-        validators.push('maxbytes="' + field.fieldValidateRulesMaxbytes +'"');
-    }
-    if (rules.indexOf('pattern') !== -1) {
-        validators.push('pattern="' + field.fieldValidateRulesPattern.replace(/"/g, '&#34;') + '"');
-    }
-
-    result = validators.join(' ');
-}
--%>
-<%- result -%>
