--- generators/entity-client4/templates/ionic/src/app/pages/entities/pagination-template.ejs
@@ -1,143 +1,0 @@
-<%#
- Copyright 2019-Present the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see http://www.jhipster.tech/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-<%_
-const i18nToLoad = [entityInstance];
-for (const idx in fields) {
-    if (fields[idx].fieldIsEnum === true) {
-        i18nToLoad.push(fields[idx].enumInstance);
-    }
-}
-_%>
-
-currentAccount: any;
-    <%= entityInstancePlural %>: <%= entityAngularName %>[];
-    error: any;
-    success: any;
-    eventSubscriber: Subscription;
-    <%_ if (searchEngine === 'elasticsearch') { _%>
-    currentSearch: string;
-    <%_ } _%>
-    <%_ if (databaseType !== 'cassandra') { _%>
-    routeData: any;
-    links: any;
-    totalItems: any;
-    queryCount: any;
-    itemsPerPage: any;
-    page: any;
-    predicate: any;
-    previousPage: any;
-    reverse: any;
-    <%_ } _%>
-
-    constructor(
-        private <%= entityInstance %>Service: <%= entityAngularName %>Service,
-        private parseLinks: JhiParseLinks,
-        private jhiAlertService: JhiAlertService,
-        private principal: Principal,
-        private activatedRoute: ActivatedRoute,
-        <%_ if (fieldsContainBlob) { _%>
-        private dataUtils: JhiDataUtils,
-        <%_ } _%>
-        private router: Router,
-        private eventManager: JhiEventManager
-    ) {
-        <%_ if (databaseType !== 'cassandra') { _%>
-        this.itemsPerPage = ITEMS_PER_PAGE;
-        this.routeData = this.activatedRoute.data.subscribe((data) => {
-            this.page = data['pagingParams'].page;
-            this.previousPage = data['pagingParams'].page;
-            this.reverse = data['pagingParams'].ascending;
-            this.predicate = data['pagingParams'].predicate;
-        });
-        <%_ if (searchEngine === 'elasticsearch') { _%>
-        this.currentSearch = activatedRoute.snapshot.params['search'] ? activatedRoute.snapshot.params['search'] : '';
-        <%_ } _%>
-        <%_ } _%>
-    }
-
-    loadAll() {
-        <%_ if (searchEngine === 'elasticsearch') { _%>
-        if (this.currentSearch) {
-            this.<%= entityInstance %>Service.search({
-                page: this.page - 1,
-                query: this.currentSearch,
-                size: this.itemsPerPage,
-                sort: this.sort()}).subscribe(
-                    (res: ResponseWrapper) => this.onSuccess(res.json, res.headers),
-                    (res: ResponseWrapper) => this.onError(res.json)
-                );
-            return;
-        }
-        <%_ } _%>
-        this.<%= entityInstance %>Service.query(<%_ if (databaseType !== 'cassandra') { _%>{
-            page: this.page - 1,
-            size: this.itemsPerPage,
-            sort: this.sort()}<%_ } _%>).subscribe(
-            (res: ResponseWrapper) => this.onSuccess(res.json, res.headers),
-            (res: ResponseWrapper) => this.onError(res.json)
-        );
-    }
-    <%_ if (databaseType !== 'cassandra') { _%>
-    loadPage(page: number) {
-        if (page !== this.previousPage) {
-            this.previousPage = page;
-            this.transition();
-        }
-    }
-    transition() {
-        this.router.navigate(['/<%= entityUrl %>'], {queryParams:
-            {
-                page: this.page,
-                size: this.itemsPerPage,
-                <%_ if (searchEngine === 'elasticsearch') { _%>
-                search: this.currentSearch,
-                <%_ } _%>
-                sort: this.predicate + ',' + (this.reverse ? 'asc' : 'desc')
-            }
-        });
-        this.loadAll();
-    }
-
-    clear() {
-        this.page = 0;
-        <%_ if(searchEngine === 'elasticsearch') { _%>
-        this.currentSearch = '';
-        <%_ } _%>
-        this.router.navigate(['/<%= entityUrl %>', {
-            page: this.page,
-            sort: this.predicate + ',' + (this.reverse ? 'asc' : 'desc')
-        }]);
-        this.loadAll();
-    }
-    <%_ } _%>
-    <%_ if(searchEngine === 'elasticsearch') { _%>
-    search(query) {
-        if (!query) {
-            return this.clear();
-        }
-        this.page = 0;
-        this.currentSearch = query;
-        this.router.navigate(['/<%= entityUrl %>', {
-            search: this.currentSearch,
-            page: this.page,
-            sort: this.predicate + ',' + (this.reverse ? 'asc' : 'desc')
-        }]);
-        this.loadAll();
-    }
-    <%_ } _%>
