+++ test-integration/install/01-install-jhipster-stack.sh
@@ -15,17 +15,16 @@ echo $IONIC4J_BRANCH
 # Clone JHipster
 #-------------------------------------------------------------------------------
 cd "$HOME"
-npm i -g generator-jhipster@6.10.5
-#git clone $JHIPSTER_REPO generator-jhipster
-#cd generator-jhipster
-#if [ "$JHIPSTER_BRANCH" == "latest" ]; then
-#    LATEST=$(git describe --abbrev=0)
-#    git checkout -b $LATEST $LATEST
-#elif [ "$JHIPSTER_BRANCH" != "main" ]; then
-#    git checkout -b $JHIPSTER_BRANCH origin/$JHIPSTER_BRANCH
-#fi
-#git --no-pager log -n 10 --graph --pretty='%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit
-#npm link
+git clone $JHIPSTER_REPO generator-jhipster
+cd generator-jhipster
+if [ "$JHIPSTER_BRANCH" == "latest" ]; then
+    LATEST=$(git describe --abbrev=0)
+    git checkout -b $LATEST $LATEST
+elif [ "$JHIPSTER_BRANCH" != "main" ]; then
+    git checkout -b $JHIPSTER_BRANCH origin/$JHIPSTER_BRANCH
+fi
+git --no-pager log -n 10 --graph --pretty='%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit
+npm link
 
 #-------------------------------------------------------------------------------
 # Link Ionic
